apps/compliance_annotator/src/requirements.txt
pytz==2014.10
requests==2.2.1
kafka-python
elasticsearch
elasticsearch-dsl
pykafka==1.1.1